integrated security

+++ inc/deleteFromProductFromCart.php
@@ -1,5 +1,10 @@
 <?php
 require_once '../config/config.php';
+if (!isset($_SESSION['user_id'])) {
+    header('Location: ../login.php');
+    exit;
+}
+
 
 $query = "DELETE FROM `cart_items` WHERE product_id = :pid";
